+++ internal/app/fileio/storage.go
@@ -25,8 +25,8 @@ type FsFile interface {
 //go:generate go run github.com/vektra/mockery/v2@v2.52.3 --name=StorageController --filename=mock_storage.go
 type StorageController interface {
 	FileSystem
-	AllocateStorage(size int) error
-	ReleaseStorage(size int) error
+	AllocateStorage(size int64) error
+	ReleaseStorage(size int64) error
 	AllocateAll() (n int, err error)
 	AddFile(id uuid.UUID) (File, error)
 	DeleteFile(id uuid.UUID) error
